requirements.txt
attrs==19.3.0
babel==2.6.0
beautifulsoup4==4.8.2
GitPython==2.1.11
livereload==2.6.0
mako==1.0.9
pymorphy2==0.8
requests==2.21.0
sortedcontainers==2.1.0
tdh-tcd==2.4.3
cached-property==1.5.1
youtube-dl